move logger classes to new Logger namespace
also make them use file-scoped namespace
and get rid of the extra indent
