+++ Examples/Executor/Program.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Runtime.ConstrainedExecution;
 using System.Text;
 using QuickFix;
+using QuickFix.Logger;
 
 namespace Executor
 {
