+++ Examples/SimpleAcceptor/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using QuickFix;
+using QuickFix.Logger;
 using QuickFix.Transport;
 
 namespace SimpleAcceptor
