+++ QuickFIXn/ClientHandlerThread.cs
@@ -2,6 +2,7 @@
 using System.Net.Sockets;
 using System.Threading;
 using System;
+using QuickFix.Logger;
 
 namespace QuickFix
 {
