+++ QuickFIXn/Logger/NullLogFactory.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+
+namespace QuickFix.Logger;
+
+public class NullLogFactory : ILogFactory
+{
+    public NullLogFactory() { }
+
+    public ILog Create(SessionID _x)
+    {
+        return new NullLog();
+    }
+}
