--- QuickFIXn/NullLogFactory.cs
@@ -1,14 +1,0 @@
-﻿#nullable enable
-
-namespace QuickFix
-{
-    public class NullLogFactory : ILogFactory
-    {
-        public NullLogFactory() { }
-
-        public ILog Create(SessionID _x)
-        {
-            return new NullLog();
-        }
-    }
-}
