+++ QuickFIXn/Session.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Threading;
 using QuickFix.Fields;
 using QuickFix.Fields.Converters;
+using QuickFix.Logger;
 
 namespace QuickFix
 {
