+++ QuickFIXn/SessionFactory.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 using System;
 using System.Collections.Generic;
+using QuickFix.Logger;
 using QuickFix.Util;
 
 namespace QuickFix
