+++ QuickFIXn/Store/FileStoreFactory.cs
