+++ QuickFIXn/Store/IMessageStore.cs
