+++ QuickFIXn/Store/IMessageStoreFactory.cs
