+++ QuickFIXn/Store/MemoryStore.cs
