+++ QuickFIXn/Store/MemoryStoreFactory.cs
