+++ QuickFIXn/ThreadedSocketAcceptor.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System;
+using QuickFix.Logger;
 
 namespace QuickFix
 {
