+++ QuickFIXn/Transport/SocketInitiator.cs
@@ -6,6 +6,7 @@ using System.Net.Sockets;
 using System.Net;
 using System.Threading;
 using System.IO;
+using QuickFix.Logger;
 
 namespace QuickFix.Transport
 {
