+++ UnitTests/SessionStateTest.cs
@@ -5,6 +5,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
+using QuickFix.Logger;
 
 namespace UnitTests
 {
